Lazy load prism to save another 1mb
